10-15 13:44, add button bsp

+++ applications/main.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-    uint16_t count = 1;
+    int count = 1;
     /* set LED0 pin mode to output */
     rt_pin_mode(LED0_PIN, PIN_MODE_OUTPUT);
